remove dead classes
