--- Grpc/Services/GrocerService.cs
@@ -1,19 +1,0 @@
-//namespace Grpc.Services;
-
-//public class GrocerService : HubStoreService.HubStoreServiceBase
-//{
-//    public GrocerService()
-//    {
-
-//    }
-
-//    public Task<HealthCheckResponse> CheckIn(Store store) {
-
-//        return Task.FromResult(new HealthCheckResponse {
-//            StoreId = store.Id,
-            
-//        });
-
-//    }
-    
-//}
